+++ packages/ui/voo_navigation/lib/src/presentation/organisms/voo_mobile_scaffold.dart
@@ -98,6 +98,7 @@ class VooMobileScaffold extends StatelessWidget {
                   VooMobileAppBar(
                     config: config,
                     selectedItem: selectedItem,
+                    selectedId: selectedId,
                     showMenuButton: false,
                   ))
               : null,